+++ lib/src/analyzers/unused_code_analyzer/reporters/reporters_list/console/unused_code_console_reporter.dart
@@ -18,7 +18,7 @@ class UnusedCodeConsoleReporter extends ConsoleReporter<UnusedCodeFileReport,
     UnusedCodeReportParams? additionalParams,
   }) async {
     if (records.isEmpty) {
-      if (additionalParams?.congratulate ?? false) {
+      if (additionalParams?.congratulate ?? true) {
         output.writeln('${okPen('✔')} no unused code found!');
       }