+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* feat: add new command flag `--[no-]congratulate`.
+* feat: add new command flag `--no-congratulate`.
 * chore: restrict `analyzer` version to `>=2.4.0 <3.3.0`.
 * fix: make `check-unused-l10n` also cover supertype member calls.